pyproject.toml
[project]
name = "pypy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.64.0",
    "google-generativeai>=0.8.5",
    "instructor>=1.13.0",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "pandas>=2.3.3",
    "pydantic-ai>=1.27.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "tantivy>=0.25.1",
    "transformers>=4.57.3",
]
